+++ gitlab/connection_config.go
@@ -1,8 +1,8 @@
 package gitlab
 
 import (
-	"github.com/turbot/steampipe-plugin-sdk/plugin"
-	"github.com/turbot/steampipe-plugin-sdk/plugin/schema"
+	"github.com/turbot/steampipe-plugin-sdk/v3/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v3/plugin/schema"
 )
 
 type GitLabConfig struct {